FE: - added validations for login, register and adding a new movie forms - implemented CRUD on Movie entity - changing urls using route
BE:
- deleted all validations in backend (all validations will be on in FE)
- changed Movie and User queries
- added ResponseEntities so the errors can be caught on FE

+++ src/main/java/com/movie/moviesite/service/MovieService.java
@@ -3,6 +3,7 @@ package com.movie.moviesite.service;
 import com.movie.moviesite.model.Movie;
 import com.movie.moviesite.repository.MovieRepostitory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
@@ -21,19 +22,22 @@ public class MovieService {
         return this.movieRepostitory.getAllMovies();
     }
 
-    public Collection<Movie> getReviewedMovies() {
-        return this.movieRepostitory.getReviewedMovies();
-    }
-
     public Movie getMovieById(Long movieId) {
-        return this.movieRepostitory.getMovieById(movieId);
+        return this.movieRepostitory.getSimpleMovieById(movieId);
     }
 
-    public void saveMovie(Movie movie) {
-        this.movieRepostitory.save(movie);
+    public ResponseEntity<?> saveMovie(Movie movie, String directedBy) {
+        Movie oldMovie = this.movieRepostitory.getMovieByName(movie.getName());
+        if (oldMovie == null) {
+            this.movieRepostitory.save(movie);
+            this.movieRepostitory.createDirectedRel(movie.getName(), directedBy);
+            return ResponseEntity.ok(movie);
+        } else {
+            return ResponseEntity.status(409).body("This movies already exists");
+        }
     }
 
-    public void updateMovieById(Movie movie, Long movieId) {
+    public ResponseEntity<?> updateMovieById(Movie movie, Long movieId) {
         Optional<Movie> foundMovie = this.movieRepostitory.findById(movieId);
         if (foundMovie.isPresent()) {
             foundMovie.get().setId(movie.getId());
@@ -41,12 +45,20 @@ public class MovieService {
             foundMovie.get().setName(movie.getName());
             foundMovie.get().setReleaseYear(movie.getReleaseYear());
             foundMovie.get().setDirector(movie.getDirector());
-            foundMovie.get().setActedInActors(movie.getActedInActors());
             this.movieRepostitory.save(foundMovie.get());
+            return ResponseEntity.ok().body("The movie \"" + movie.getName() + "\" was updated");
+        } else {
+            return ResponseEntity.status(404).body("Movie not found");
         }
     }
 
-    public void deleteMovieById(Long movieId) {
-        this.movieRepostitory.deleteById(movieId);
+    public ResponseEntity<?> deleteMovieById(Long movieId) {
+        Movie m = this.movieRepostitory.getSimpleMovieById(movieId);
+        if (m == null) {
+            return ResponseEntity.status(404).body("Movie doesn't exist with the id " + movieId);
+        } else {
+            this.movieRepostitory.deleteById(movieId);
+            return ResponseEntity.ok().body("Deleted successfully");
+        }
     }
 }
